made an item component
So the component is ready,  but I use it in a wrong way. The page creates when you click on the item, but needs to create when item is being added to the database. As I dont have a database yet, its all good, as long as the component works. Next step is creating an admin page that will have all the functionality for working with items(adding, updating, deleting products). After creating the admin page will start working on a database.